+++ _draw/draw_sphere.c
